Move hex/bytes conversion methods up
These methods will be used in the transact signing module.
